Show only month and year in date displays
VenuePage and the Index card both rendered exact dates
('21. April 2026' / 'Apr 2026'). The day adds noise — these
are espresso visits, not appointments — so trim to month +
year everywhere user-facing:

  VenuePage  → 'April 2026' / 'April 2026'
  IndexPanel → 'April 2026' / 'April 2026' (was short month)

BottomSheet doesn't render the date so nothing to change
there. The full timestamp stays in the database for
sortability.

+++ src/components/IndexPanel.jsx
@@ -297,7 +297,7 @@ export default function IndexPanel({ venues, isOpen, country = '', setCountry = 
           const score   = venue.avg_score != null ? parseFloat(venue.avg_score) : null;
           const color   = bucketTextColor(score);
           const dateStr = venue.rated_at
-            ? new Date(venue.rated_at).toLocaleDateString(lang === 'de' ? 'de-DE' : 'en-GB', { month: 'short', year: 'numeric' })
+            ? new Date(venue.rated_at).toLocaleDateString(lang === 'de' ? 'de-DE' : 'en-GB', { month: 'long', year: 'numeric' })
             : '';
           return (
             <button
